Tests for fn folder
